+++ device-plugin/pkg/cndev/bindings.go
@@ -20,6 +20,13 @@ import "C"
 
 import (
 	"fmt"
+	"io"
+	"os"
+	"path"
+	"path/filepath"
+	"runtime"
+	"strconv"
+	"strings"
 	"time"
 	"unsafe"
 
@@ -71,11 +78,10 @@ type SmluInfo struct {
 }
 
 func Init(healthCheck bool) error {
-	r := dl.cndevInit()
-	if err := errorString(r); err != nil || healthCheck {
-		return err
+	if healthCheck {
+		return (errorString(C.cndevInit(C.int(0))))
 	}
-	return generateDeviceHandleMap()
+	return errorString(dl.cndevInit())
 }
 
 func Release() error {
@@ -83,6 +89,10 @@ func Release() error {
 }
 
 func CreateSmluProfile(pl *DsmluProfile, memUnit int) (uint, error) {
+	if ret := dl.checkExist("cndevCreateSMluProfileInfo"); ret != C.CNDEV_SUCCESS {
+		return 0, errorString(ret)
+	}
+
 	var profileID C.int
 	var profile C.cndevSMluSet_t
 	profile.version = C.CNDEV_VERSION_6
@@ -93,6 +103,10 @@ func CreateSmluProfile(pl *DsmluProfile, memUnit int) (uint, error) {
 }
 
 func CreateSmluProfileInstance(profileID, index uint) (int, error) {
+	if ret := dl.checkExist("cndevCreateSMluInstanceByProfileId"); ret != C.CNDEV_SUCCESS {
+		return 0, errorString(ret)
+	}
+
 	var instance C.cndevMluInstance_t
 	name := C.CString("")
 	defer C.free(unsafe.Pointer(name))
@@ -101,14 +115,26 @@ func CreateSmluProfileInstance(profileID, index uint) (int, error) {
 }
 
 func DestroySmlu(instanceHandle int) error {
+	if ret := dl.checkExist("cndevDestroySMluInstanceByHandle"); ret != C.CNDEV_SUCCESS {
+		return errorString(ret)
+	}
+
 	return errorString(C.cndevDestroySMluInstanceByHandle(C.int(instanceHandle)))
 }
 
 func DestroySmluProfile(profileID, index uint) error {
+	if ret := dl.checkExist("cndevDestroySMluProfileInfo"); ret != C.CNDEV_SUCCESS {
+		return errorString(ret)
+	}
+
 	return errorString(C.cndevDestroySMluProfileInfo(C.int(profileID), cndevHandleMap[index]))
 }
 
 func DeviceMimModeEnabled(idx uint) (bool, error) {
+	if ret := dl.checkExist("cndevGetMimMode"); ret != C.CNDEV_SUCCESS {
+		return false, errorString(ret)
+	}
+
 	var mode C.cndevMimMode_t
 	mode.version = C.CNDEV_VERSION_6
 	r := C.cndevGetMimMode(&mode, cndevHandleMap[idx])
@@ -116,6 +142,10 @@ func DeviceMimModeEnabled(idx uint) (bool, error) {
 }
 
 func DeviceSmluModeEnabled(idx uint) (bool, error) {
+	if ret := dl.checkExist("cndevGetSMLUMode"); ret != C.CNDEV_SUCCESS {
+		return false, errorString(ret)
+	}
+
 	var mode C.cndevSMLUMode_t
 	mode.version = C.CNDEV_VERSION_6
 	r := C.cndevGetSMLUMode(&mode, cndevHandleMap[idx])
@@ -123,6 +153,10 @@ func DeviceSmluModeEnabled(idx uint) (bool, error) {
 }
 
 func GetAllMluInstanceInfo(idx uint) ([]MimInfo, error) {
+	if ret := dl.checkExist("cndevGetAllMluInstanceInfo"); ret != C.CNDEV_SUCCESS {
+		return nil, errorString(ret)
+	}
+
 	miCount := C.int(1 << 4)
 	var miInfos []MimInfo
 	var miInfo C.cndevMluInstanceInfo_t
@@ -172,6 +206,10 @@ func GetAllMluInstanceInfo(idx uint) ([]MimInfo, error) {
 }
 
 func GetAllSmluInfo(idx uint) ([]SmluInfo, error) {
+	if ret := dl.checkExist("cndevGetAllSMluInstanceInfo"); ret != C.CNDEV_SUCCESS {
+		return nil, errorString(ret)
+	}
+
 	smluCount := C.int(1 << 7)
 	var smluInfos []SmluInfo
 	var smluInfo C.cndevSMluInfo_t
@@ -222,6 +260,10 @@ func GetAllSmluInfo(idx uint) ([]SmluInfo, error) {
 }
 
 func GetDeviceCount() (uint, error) {
+	if ret := dl.checkExist("cndevGetDeviceCount"); ret != C.CNDEV_SUCCESS {
+		return 0, errorString(ret)
+	}
+
 	var cardInfos C.cndevCardInfo_t
 	cardInfos.version = C.CNDEV_VERSION_6
 	r := C.cndevGetDeviceCount(&cardInfos)
@@ -229,16 +271,28 @@ func GetDeviceCount() (uint, error) {
 }
 
 func GetDeviceMemory(idx uint) (uint, error) {
+	if ret := dl.checkExist("cndevGetMemoryUsageV2"); ret != C.CNDEV_SUCCESS {
+		return 0, errorString(ret)
+	}
+
 	var cardMemInfo C.cndevMemoryInfoV2_t
 	r := C.cndevGetMemoryUsageV2(&cardMemInfo, cndevHandleMap[idx])
 	return uint(cardMemInfo.physicalMemoryTotal), errorString(r)
 }
 
 func GetDeviceModel(idx uint) string {
+	if ret := dl.checkExist("cndevGetCardNameStringByDevId"); ret != C.CNDEV_SUCCESS {
+		return ""
+	}
+
 	return C.GoString(C.cndevGetCardNameStringByDevId(cndevHandleMap[idx]))
 }
 
 func GetDeviceProfileInfo(index uint) ([]DsmluProfileInfo, error) {
+	if ret := dl.checkExist("cndevGetSMluProfileIdInfo"); ret != C.CNDEV_SUCCESS {
+		return nil, errorString(ret)
+	}
+
 	var dsmluProfileInfos []DsmluProfileInfo
 	var deviceProfiles C.cndevSMluProfileIdInfo_t
 	deviceProfiles.version = C.CNDEV_VERSION_6
@@ -268,6 +322,10 @@ func GetDeviceProfileInfo(index uint) ([]DsmluProfileInfo, error) {
 }
 
 func GetDeviceUUID(idx uint) (string, error) {
+	if ret := dl.checkExist("cndevGetUUID"); ret != C.CNDEV_SUCCESS {
+		return "", errorString(ret)
+	}
+
 	var uuidInfo C.cndevUUID_t
 	uuidInfo.version = C.CNDEV_VERSION_6
 	r := C.cndevGetUUID(&uuidInfo, cndevHandleMap[idx])
@@ -345,6 +403,10 @@ func GetMLULinkGroups() ([][]uint, error) {
 }
 
 func GetSmluInfo(instanceHandle int) (SmluInfo, error) {
+	if ret := dl.checkExist("cndevGetSMluInstanceInfo"); ret != C.CNDEV_SUCCESS {
+		return SmluInfo{}, errorString(ret)
+	}
+
 	var smluInfo C.cndevSMluInfo_t
 	smluInfo.version = C.CNDEV_VERSION_6
 	r := C.cndevGetSMluInstanceInfo(&smluInfo, C.int(instanceHandle))
@@ -378,6 +440,10 @@ func NewDeviceLite(idx uint) (*Device, error) {
 }
 
 func errorString(ret C.cndevRet_t) error {
+	if r := dl.checkExist("cndevGetErrorString"); r != C.CNDEV_SUCCESS {
+		return errorString(r)
+	}
+
 	if ret == C.CNDEV_SUCCESS {
 		return nil
 	}
@@ -386,6 +452,10 @@ func errorString(ret C.cndevRet_t) error {
 }
 
 func GetDeviceComputeMode(idx uint, delayTime int) (bool, error) {
+	if ret := dl.checkExist("cndevGetComputeMode"); ret != C.CNDEV_SUCCESS {
+		return false, errorString(ret)
+	}
+
 	var ret C.cndevRet_t
 	var cardComputeMode C.cndevComputeMode_t
 	cardComputeMode.version = C.CNDEV_VERSION_6
@@ -396,6 +466,10 @@ func GetDeviceComputeMode(idx uint, delayTime int) (bool, error) {
 }
 
 func getDeviceInfo(idx uint) (string, string, string, string, error) {
+	if ret := dl.checkExist("cndevGetCardName", "cndevGetCardSN", "cndevGetUUID"); ret != C.CNDEV_SUCCESS {
+		return "", "", "", "", errorString(ret)
+	}
+
 	var cardName C.cndevCardName_t
 	var cardSN C.cndevCardSN_t
 	var uuidInfo C.cndevUUID_t
@@ -431,6 +505,10 @@ func getDeviceInfo(idx uint) (string, string, string, string, error) {
 }
 
 func GetDeviceHealthState(idx uint, delayTime int) (int, bool, bool, error) {
+	if ret := dl.checkExist("cndevGetCardHealthState"); ret != C.CNDEV_SUCCESS {
+		return 0, false, false, errorString(ret)
+	}
+
 	var ret C.cndevRet_t
 	var cardHealthState C.cndevCardHealthState_t
 	var healthCode int
@@ -443,6 +521,10 @@ func GetDeviceHealthState(idx uint, delayTime int) (int, bool, bool, error) {
 }
 
 func getDeviceMLULinkDevs(idx uint) (map[string]int, error) {
+	if ret := dl.checkExist("cndevGetMLULinkPortNumber", "cndevGetMLULinkStatusV2", "cndevGetMLULinkRemoteInfo"); ret != C.CNDEV_SUCCESS {
+		return nil, errorString(ret)
+	}
+
 	devs := make(map[string]int)
 	portNum := C.cndevGetMLULinkPortNumber(cndevHandleMap[idx])
 	for i := 0; i < int(portNum); i++ {
@@ -471,19 +553,23 @@ func getDeviceMLULinkDevs(idx uint) (map[string]int, error) {
 }
 
 func getDeviceNUMA(idx uint) (int, error) {
+	if ret := dl.checkExist("cndevGetNUMANodeIdByDevId"); ret != C.CNDEV_SUCCESS {
+		return 0, errorString(ret)
+	}
+
 	var numaNode C.cndevNUMANodeId_t
 	numaNode.version = C.CNDEV_VERSION_6
 	r := C.cndevGetNUMANodeIdByDevId(&numaNode, cndevHandleMap[idx])
 	return int(numaNode.nodeId), errorString(r)
 }
 
-func generateDeviceHandleMap() error {
-	num, err := GetDeviceCount()
-	if err != nil {
-		return err
+func generateDeviceHandleMap(count uint) error {
+	if ret := dl.checkExist("cndevGetDeviceHandleByIndex"); ret != C.CNDEV_SUCCESS {
+		return errorString(ret)
 	}
+
 	cndevHandleMap = map[uint]C.cndevDevice_t{}
-	for i := uint(0); i < num; i++ {
+	for i := uint(0); i < count; i++ {
 		var handle C.cndevDevice_t
 		r := C.cndevGetDeviceHandleByIndex(C.int(i), &handle)
 		if errorString(r) != nil {
@@ -492,4 +578,180 @@ func generateDeviceHandleMap() error {
 		cndevHandleMap[i] = handle
 	}
 	return nil
+}
+
+func FetchMLUCounts() (uint, error) {
+	targetVendorID := uint16(0xcabc) // cambricon mlu vendor ID is 0xcabc
+	targetClassBase := uint8(0x12)   // cambricon mlu class code base is 0x12
+	pciDevicesPath := "/sys/bus/pci/devices"
+
+	readHexFile := func(path string) (uint64, error) {
+		data, err := os.ReadFile(path)
+		if err != nil {
+			return 0, err
+		}
+		s := strings.TrimSpace(string(data))
+		s = strings.TrimPrefix(s, "0x")
+		val, err := strconv.ParseUint(s, 16, 32)
+		if err != nil {
+			return 0, err
+		}
+		return val, nil
+	}
+
+	entries, err := os.ReadDir(pciDevicesPath)
+	if err != nil {
+		log.Errorf("Can't read pci dir: %v", err)
+		return 0, err
+	}
+
+	var count uint
+	for _, entry := range entries {
+		devicePath := filepath.Join(pciDevicesPath, entry.Name())
+		vendorID, err := readHexFile(filepath.Join(devicePath, "vendor"))
+		if err != nil {
+			log.Warnf("Can't read vendor file: %v", err)
+			continue
+		}
+		if uint16(vendorID) != targetVendorID {
+			log.Debugf("VendorID not match 0x%x", vendorID)
+			continue
+		}
+		classCode, err := readHexFile(filepath.Join(devicePath, "class"))
+		if err != nil {
+			log.Warnf("Can't read class file: %v", err)
+			continue
+		}
+		classBase := uint8((classCode >> 16) & 0xFF)
+		if classBase == targetClassBase {
+			log.Debugf("Find mlu device: %s with vendorID 0x%x,classCode: 0x%x, classBase: 0x%x", devicePath, vendorID, classCode, classBase)
+			count++
+		}
+	}
+
+	log.Debugf("Find %d mlu devices", count)
+	return count, nil
+}
+
+func isDriverRunning(counts uint) bool {
+	for i := range counts {
+		_, good, running, err := GetDeviceHealthState(i, 0)
+		if err != nil {
+			log.Warnf("GetDeviceHealth for slot %d with err %v", i, err)
+			return false
+		}
+		if !good {
+			log.Warnf("MLU device %d health maybe in problem, ignoring at init", i)
+		}
+		if !running {
+			log.Warnf("MLU device %d driver is not running", i)
+			return false
+		}
+	}
+	return true
+}
+
+func EnsureMLUAllOk() {
+	log.Infof("Start to ensure mlu driver status is ok")
+	i := 1
+	for {
+		if i < 60000 {
+			i = i << 1
+		}
+		time.Sleep(time.Duration(min(i+100, 60000)) * time.Millisecond)
+
+		if i == 2 {
+			if err := Init(false); err != nil {
+				log.Errorf("Init cndev client failed with err: %v", err)
+				continue
+			}
+		} else {
+			if err := Init(true); err != nil {
+				log.Errorf("Init cndev client failed with err: %v", err)
+				continue
+			}
+		}
+
+		if os.Getenv("INTEGRATION_TEST") == "true" {
+			counts, _ := GetDeviceCount()
+			generateDeviceHandleMap(counts)
+			log.Info("Skip ensuring mlu driver status is ok for integration test")
+			return
+		}
+
+		log.Debug("Start GetDeviceCount")
+		counts, err := GetDeviceCount()
+		if err != nil {
+			log.Errorf("GetDeviceCount failed %v", err)
+			continue
+		}
+		if counts == 0 {
+			log.Warn("No MLU device found with GetDeviceCount")
+			continue
+		}
+		log.Debugf("Device counts: %d", counts)
+
+		realCounts, err := FetchMLUCounts()
+		if err != nil {
+			log.Errorf("FetchMLUCounts failed %v", err)
+			continue
+		}
+		log.Debugf("RealCounts is :%d ", realCounts)
+
+		if counts != realCounts {
+			log.Warnf("MLU device count not match, counts: %d, realCounts: %d", counts, realCounts)
+			continue
+		}
+		log.Infof("MLU device count match, count is %d", counts)
+
+		if err := generateDeviceHandleMap(counts); err != nil {
+			log.Panicf("Generate Device Handle Map failed, this should never happen, counts: %d, err: %v", counts, err)
+		}
+
+		if !isDriverRunning(counts) {
+			continue
+		}
+
+		log.Info("Driver of MLU devices are all running")
+		return
+	}
+}
+
+func EnsureCndevLib() error {
+	arch := runtime.GOARCH
+	log.Infof("Ensuring cndev lib, arch is %s", arch)
+	var targetDir string
+	switch arch {
+	case "386", "amd64":
+		targetDir = "x86_64-linux-gnu"
+	case "arm", "arm64":
+		targetDir = "aarch64-linux-gnu"
+	default:
+		log.Infof("Invalid arch %s", arch)
+		return nil
+	}
+
+	src := path.Join("/host/usr/lib", targetDir, "libcndev.so")
+	dst := "/usr/lib/libcndev.so"
+	if _, err := os.Stat(src); err != nil {
+		log.Info("Found no libcndev.so on host, use default")
+		return nil
+	}
+	log.Info("Found libcndev.so in host, try to copy to /usr/lib")
+	sourceFile, err := os.Open(src)
+	if err != nil {
+		log.Errorf("Can't open %s", src)
+		return err
+	}
+	defer sourceFile.Close()
+
+	destinationFile, err := os.Create(dst)
+	if err != nil {
+		log.Errorf("Can't create %s", dst)
+		return err
+	}
+	defer destinationFile.Close()
+
+	_, err = io.Copy(destinationFile, sourceFile)
+	return err
 }